requirements.txt
flask==2.3.3
huggingface-hub>=0.20.0
google-generativeai==0.3.2
pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
